move things out of paneview.ts
